refactor: moved test utility function from desmosclient.integration.spec.ts
